tests-adjusted-to-new-UI
