+++ ocs_ci/ocs/ui/page_objects/backing_store_tab.py
@@ -5,9 +5,10 @@ from ocs_ci.ocs.ui.page_objects.data_foundation_tabs_common import (
     DataFoundationDefaultTab,
     CreateResourceForm,
 )
+from ocs_ci.ocs.ui.page_objects.object_service import ObjectService
 
 
-class BackingStoreTab(DataFoundationDefaultTab, CreateResourceForm):
+class BackingStoreTab(ObjectService, CreateResourceForm):
     def __init__(self):
         DataFoundationDefaultTab.__init__(self)
         CreateResourceForm.__init__(self)
