+++ ocs_ci/ocs/ui/page_objects/bucket_class_tab.py
@@ -7,13 +7,13 @@ from ocs_ci.ocs import constants
 from ocs_ci.ocs.ocp import OCP
 from ocs_ci.ocs.ui.base_ui import logger
 from ocs_ci.ocs.ui.page_objects.data_foundation_tabs_common import (
-    DataFoundationDefaultTab,
     CreateResourceForm,
     DataFoundationTabBar,
 )
+from ocs_ci.ocs.ui.page_objects.object_service import ObjectService
 
 
-class BucketClassTab(DataFoundationDefaultTab, CreateResourceForm):
+class BucketClassTab(ObjectService, CreateResourceForm):
     def __init__(self):
         DataFoundationTabBar.__init__(self)
         CreateResourceForm.__init__(self)
