+++ ocs_ci/ocs/ui/views.py
@@ -487,8 +487,8 @@ page_nav = {
     "volumesnapshots_page": ("VolumeSnapshots", By.LINK_TEXT),
     "volumesnapshotclasses_page": ("VolumeSnapshotClasses", By.LINK_TEXT),
     "volumesnapshotcontents_page": ("VolumeSnapshotContents", By.LINK_TEXT),
-    "object_buckets_page": ("Object Buckets", By.LINK_TEXT),
-    "object_bucket_claims_page": ("Object Bucket Claims", By.LINK_TEXT),
+    "object_buckets_tab": ("//a[normalize-space()='Object Buckets']", By.XPATH),
+    "object_storage": ("//a[normalize-space()='Object Storage']", By.XPATH),
     "Monitoring": ("//button[text()='Monitoring']", By.XPATH),
     "alerting_page": ("Alerting", By.LINK_TEXT),
     "metrics_page": ("Metrics", By.LINK_TEXT),
